refactor: LottoMachine 패키지 domain으로 이동
