+++ src/main/java/lotto/controller/LottoController.java
@@ -6,7 +6,6 @@ import lotto.dto.request.MoneyRequestDto;
 import lotto.dto.request.WinningNumbersRequestDto;
 import lotto.dto.response.LottoResultResponseDto;
 import lotto.dto.response.LottoStatusesResponseDto;
-import lotto.service.LottoMachine;
 import lotto.view.InputView;
 import lotto.view.OutputView;
