+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -1,8 +1,5 @@
-package lotto.service;
+package lotto.domain;
 
-import lotto.domain.Lotto;
-import lotto.domain.Lottos;
-import lotto.domain.Money;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
